almost done with mods

+++ PixelDungeon/src/main/java/com/watabou/pixeldungeon/windows/WndModSelect.java
@@ -5,6 +5,7 @@ import com.nyrds.android.util.DownloadTask;
 import com.nyrds.android.util.FileSystem;
 import com.nyrds.android.util.GuiProperties;
 import com.nyrds.android.util.ModdingMode;
+import com.nyrds.android.util.Mods;
 import com.nyrds.android.util.Unzip;
 import com.nyrds.android.util.Util;
 import com.nyrds.pixeldungeon.ml.R;
@@ -18,20 +19,12 @@ import com.watabou.pixeldungeon.ui.Icons;
 import com.watabou.pixeldungeon.ui.RedButton;
 import com.watabou.pixeldungeon.ui.SimpleButton;
 import com.watabou.pixeldungeon.ui.Window;
-import com.watabou.pixeldungeon.utils.GLog;
 import com.watabou.pixeldungeon.utils.Utils;
 
-import org.json.JSONArray;
-import org.json.JSONException;
-
 import java.io.File;
-import java.util.ArrayList;
-import java.util.HashMap;
 import java.util.Map;
 
 public class WndModSelect extends Window implements DownloadStateListener {
-
-	private static ArrayList<String> mMods = new ArrayList<>();
 
 	private static final int WIDTH         = 120;
 	private static final int MARGIN        = 2;
@@ -39,15 +32,10 @@ public class WndModSelect extends Window implements DownloadStateListener {
 
 	private Text downloadProgress;
 
-	static private class ModDesc {
-		public String name;
-		public String link;
-	}
-
 	private String selectedMod;
 	private String downloadTo;
 
-	private static Map<String, ModDesc> mModsMap = new HashMap<>();
+	private Map<String, Mods.ModDesc> modsList = Mods.buildModsList();
 
 	public WndModSelect() {
 		Text tfTitle = PixelScene.createMultiline(Game.getVar(R.string.ModsButton_SelectMod), GuiProperties.titleFontSize());
@@ -59,19 +47,16 @@ public class WndModSelect extends Window implements DownloadStateListener {
 
 		float pos = tfTitle.y + tfTitle.height() + MARGIN;
 
-		ArrayList<String> options = buildModsList();
-
-		for (int i = 0; i < options.size(); i++) {
-			final int index = i;
-			String option = options.get(index);
-
+		int index = 0;
+		for (Map.Entry<String, Mods.ModDesc> entry : modsList.entrySet()) {
+			final Mods.ModDesc desc = entry.getValue();
 			float additionalMargin = 0;
 
-			if (!option.equals(ModdingMode.REMIXED) && !option.contains("Download")) {
+			if (desc.installed) {
 
 				SimpleButton deleteBtn = new SimpleButton(Icons.get(Icons.CLOSE)) {
 					protected void onClick() {
-						onDelete(index);
+						onDelete(desc.name);
 					}
 				};
 				deleteBtn.setPos(WIDTH - deleteBtn.width() - MARGIN, pos);
@@ -79,73 +64,38 @@ public class WndModSelect extends Window implements DownloadStateListener {
 				add(deleteBtn);
 			}
 
-			RedButton btn = new RedButton(options.get(index)) {
+			String option = desc.name;
+			if(desc.needUpdate) {
+				option = "Update " + option;
+			}
+
+			RedButton btn = new RedButton(option) {
 
 				@Override
 				protected void onClick() {
 					hide();
-					onSelect(index);
+					onSelect(desc.name);
 				}
 
 			};
+
 			btn.setRect(MARGIN, pos, WIDTH - MARGIN * 2 - additionalMargin, BUTTON_HEIGHT);
 			add(btn);
 
 			pos += BUTTON_HEIGHT + MARGIN;
+
+
+			index = index + 1;
 		}
 
 		resize(WIDTH, (int) pos);
 	}
 
-	private static ArrayList<String> buildModsList() {
-		mModsMap.clear();
-		mMods.clear();
+	private void onDelete(String name) {
 
-		File[] extList = FileSystem.listExternalStorage();
-		final ArrayList<String> mods = new ArrayList<>();
+		File modDir = FileSystem.getExternalStorageFile(name);
 
-		String[] knownMods = Game.getVars(R.array.known_mods);
-
-		for (String knownMod : knownMods) {
-
-			try {
-				JSONArray modDesc = new JSONArray(knownMod);
-				ModDesc desc = new ModDesc();
-				desc.name = modDesc.getString(0);
-
-				// TODO check versions...
-				if (FileSystem.getExternalStorageFile(desc.name).exists()) {
-					continue;
-				}
-
-				desc.link = modDesc.getString(1);
-
-				String option = "Download " + desc.name;
-				mModsMap.put(option, desc);
-				mods.add(option);
-
-			} catch (JSONException e) {
-				GLog.w(e.getMessage());
-			}
-		}
-
-		mods.add(ModdingMode.REMIXED);
-
-		for (File file : extList) {
-			if (file.isDirectory()) {
-				mods.add(file.getName());
-			}
-		}
-		mMods = mods;
-		return mods;
-	}
-
-	private void onDelete(int index) {
-		String option = mMods.get(index);
-
-		File modDir = FileSystem.getExternalStorageFile(option);
-
-		if (modDir.exists()) {
+		if (modDir.exists() && modDir.isDirectory()) {
 			FileSystem.deleteRecursive(modDir);
 		}
 
@@ -156,21 +106,23 @@ public class WndModSelect extends Window implements DownloadStateListener {
 		Game.scene().add(new WndModDescription(ModdingMode.REMIXED));
 	}
 
-	protected void onSelect(int index) {
-		String option = mMods.get(index);
+	protected void onSelect(String option) {
 
-		File modDir = FileSystem.getExternalStorageFile(option);
-
-		if (!modDir.exists() && !option.equals(ModdingMode.REMIXED)) {
+		Mods.ModDesc desc = modsList.get(option);
+		if (!option.equals(ModdingMode.REMIXED)) {
 			if (!Util.isConnectedToInternet()) {
 				PixelDungeon.scene().add(new WndError("Please enable Internet access to download mods"));
 				return;
 			}
-			selectedMod = mModsMap.get(option).name;
-			downloadTo = FileSystem.getExternalStorageFile(selectedMod + ".zip").getAbsolutePath();
 
-			new DownloadTask(this).execute(mModsMap.get(option).link, downloadTo);
-			return;
+			if(desc.needUpdate) {
+				FileSystem.deleteRecursive(FileSystem.getExternalStorageFile(desc.name));
+				selectedMod = desc.name;
+				downloadTo = FileSystem.getExternalStorageFile(selectedMod + ".zip").getAbsolutePath();
+				desc.needUpdate = false;
+				new DownloadTask(this).execute(desc.url, downloadTo);
+				return;
+			}
 		}
 
 		String prevMod = PixelDungeon.activeMod();
